sn88-bot/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
numpy>=1.26.0
pandas>=2.1.0
lightgbm>=4.3.0
scikit-learn>=1.3.0
